feat: add foreign assets support on asset hubs, such as MYTH on PAH
