+++ packages/extension-polkagate/src/fullscreen/accountDetails/components/AssetSelect.tsx
@@ -14,7 +14,7 @@ interface Props {
   onChange: (value: number | string) => void;
   label: string;
   style: SxProps<Theme> | undefined;
-  assetId: number | undefined;
+  assetId: number | string | undefined;
   setAssetId: React.Dispatch<React.SetStateAction<number | undefined>>
 }
