+++ packages/extension-polkagate/src/fullscreen/sendFund/InputPage.tsx
@@ -6,6 +6,7 @@
 import type { IconProp } from '@fortawesome/fontawesome-svg-core';
 import type { SubmittableExtrinsicFunction } from '@polkadot/api/types';
 import type { Balance } from '@polkadot/types/interfaces';
+import type { HexString } from '@polkadot/util/types';
 import type { BalancesInfo, DropdownOption, TransferType } from '../../util/types';
 import type { Inputs } from '.';
 
@@ -22,7 +23,7 @@ import { useTranslation } from '../../components/translate';
 import { useInfo, useTeleport } from '../../hooks';
 import { getValue } from '../../popup/account/util';
 import { ASSET_HUBS } from '../../util/constants';
-import { amountToHuman, amountToMachine } from '../../util/utils';
+import { amountToHuman, amountToMachine, decodeMultiLocation } from '../../util/utils';
 import { openOrFocusTab } from '../accountDetails/components/CommonTasks';
 import { toTitleCase } from '../governance/utils/util';
 import { STEPS } from '../stake/pool/stake';
@@ -30,7 +31,7 @@ import { STEPS } from '../stake/pool/stake';
 interface Props {
   address: string;
   balances: BalancesInfo | undefined;
-  assetId: number | undefined;
+  assetId: string | undefined;
   inputs: Inputs | undefined;
   setStep: React.Dispatch<React.SetStateAction<number>>;
   setInputs: React.Dispatch<React.SetStateAction<Inputs | undefined>>;
@@ -99,6 +100,14 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
   const { api, chain, formatted } = useInfo(address);
   const teleportState = useTeleport(address);
 
+  const isForeignAsset = assetId && assetId.startsWith('0x');
+
+  const parsedAssetId = assetId === undefined || assetId === 'undefined'
+    ? undefined
+    : isForeignAsset
+      ? decodeMultiLocation(assetId as HexString)
+      : parseInt(assetId);
+
   const [amount, setAmount] = useState<string>(inputs?.amount || '0');
   const [estimatedFee, setEstimatedFee] = useState<Balance>();
   const [estimatedCrossChainFee, setEstimatedCrossChainFee] = useState<Balance>();
@@ -159,26 +168,34 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
       return undefined;
     }
 
-    const module = assetId !== undefined
-      ? isAssethub(chain.genesisHash)
-        ? 'assets'
-        : api.tx?.['currencies']
-          ? 'currencies'
-          : 'tokens'
-      : 'balances';
+    try {
+      const module = assetId !== undefined
+        ? isAssethub(chain.genesisHash)
+          ? isForeignAsset
+            ? 'foreignAssets'
+            : 'assets'
+          : api.tx?.['currencies']
+            ? 'currencies'
+            : 'tokens'
+        : 'balances';
 
-    if (['currencies', 'tokens'].includes(module)) {
-      return api.tx[module]['transfer'];
+      if (['currencies', 'tokens'].includes(module)) {
+        return api.tx[module]['transfer'];
+      }
+
+      return api.tx?.[module] && (
+        transferType === 'Normal'
+          ? api.tx[module]['transferKeepAlive']
+          : assetId !== undefined
+            ? api.tx[module]['transfer']
+            : api.tx[module]['transferAll']
+      );
+    } catch (e) {
+      console.log('Something wrong while making on chain call!', e);
+
+      return undefined;
     }
-
-    return api.tx?.[module] && (
-      transferType === 'Normal'
-        ? api.tx[module]['transferKeepAlive']
-        : assetId !== undefined
-          ? api.tx[module]['transfer']
-          : api.tx[module]['transferAll']
-    );
-  }, [api, assetId, chain, transferType]);
+  }, [api, assetId, chain, isForeignAsset, transferType]);
 
   const call = useMemo((): SubmittableExtrinsicFunction<'promise'> | undefined => {
     if (!api) {
@@ -221,11 +238,11 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
     const _params = assetId !== undefined
       ? ['currencies', 'tokens'].includes(onChainCall.section)
         ? [formatted, balances.currencyId, amount]
-        : [assetId, formatted, amount]
+        : [parsedAssetId, formatted, amount]
       : [formatted, amount];
 
     onChainCall(..._params).paymentInfo(formatted).then((i) => setFeeCall(i?.partialFee)).catch(console.error);
-  }, [api, formatted, balances, onChainCall, assetId]);
+  }, [api, formatted, balances, onChainCall, assetId, parsedAssetId]);
 
   const crossChainParams = useMemo(() => {
     if (!api || !balances || !teleportState || isCrossChain === false || (recipientParaId === INVALID_PARA_ID && !teleportState?.isParaTeleport) || Number(amount) === 0) {
@@ -288,7 +305,7 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
         : assetId !== undefined
           ? ['currencies', 'tokens'].includes(onChainCall?.section || '')
             ? [recipientAddress, balances.currencyId, amountAsBN] // this is for transferring on mutliasset chains
-            : [assetId, recipientAddress, amountAsBN] // this is for transferring on asset hubs
+            : [parsedAssetId, recipientAddress, amountAsBN] // this is for transferring on asset hubs
           : transferType === 'All'
             ? [recipientAddress, false] // transferAll with keepalive = false
             : [recipientAddress, amountAsBN]) as unknown[],
@@ -297,7 +314,7 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
       recipientGenesisHashOrParaId: recipientChainGenesisHash,
       totalFee: estimatedFee ? estimatedFee.add(estimatedCrossChainFee || BN_ZERO) : undefined
     });
-  }, [amountAsBN, estimatedFee, estimatedCrossChainFee, setInputs, call, recipientAddress, isCrossChain, crossChainParams, assetId, formatted, amount, recipientChainName, recipientChainGenesisHash, transferType, onChainCall?.section, balances]);
+  }, [amountAsBN, estimatedFee, estimatedCrossChainFee, setInputs, call, parsedAssetId, recipientAddress, isCrossChain, crossChainParams, assetId, formatted, amount, recipientChainName, recipientChainGenesisHash, transferType, onChainCall?.section, balances]);
 
   useEffect(() => {
     if (!api || !transferableBalance) {
@@ -339,21 +356,21 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
   }, [call, formatted, isCrossChain, crossChainParams]);
 
   const setWholeAmount = useCallback(() => {
-    if (!api || !transferableBalance || !maxFee || !balances || !ED) {
+    if (!transferableBalance || !maxFee || !balances) {
       return;
     }
 
     setTransferType('All');
 
-    const _isAvailableZero = transferableBalance.isZero();
+    const isAvailableZero = transferableBalance.isZero();
 
     const _maxFee = assetId === undefined ? maxFee : BN_ZERO;
 
-    const _canNotTransfer = _isAvailableZero || _maxFee.gte(transferableBalance);
-    const allAmount = _canNotTransfer ? '0' : amountToHuman(transferableBalance.sub(_maxFee).toString(), balances.decimal);
+    const canNotTransfer = isAvailableZero || _maxFee.gte(transferableBalance);
+    const allAmount = canNotTransfer ? '0' : amountToHuman(transferableBalance.sub(_maxFee).toString(), balances.decimal);
 
     setAmount(allAmount);
-  }, [api, assetId, balances, ED, maxFee, transferableBalance]);
+  }, [assetId, balances, maxFee, transferableBalance]);
 
   const _onChangeAmount = useCallback((value: string) => {
     if (!balances) {
@@ -403,7 +420,7 @@ export default function InputPage ({ address, assetId, balances, inputs, setInpu
           <ShowBalance api={api} balance={estimatedFee} />
         </Grid>
         <Grid alignItems='center' container item justifyContent='space-between' sx={{ height: '25px', width: '57.5%' }}>
-          <Infotip2 showInfoMark text={t('Existential Deposit: {{ED}}', { replace: { ED: api?.createType('Balance', ED)?.toHuman() } })}>
+          <Infotip2 showInfoMark text={t('Existential Deposit: {{ED}}', { replace: { ED: `${amountToHuman(balances?.ED, balances?.decimal)} ${balances?.token}` } })}>
             <Typography fontSize='14px' fontWeight={400}>
               {t('Transferable amount')}
             </Typography>
