+++ packages/extension-polkagate/src/hooks/index.ts
@@ -21,6 +21,9 @@ export { default as useAssetsBalances } from './useAssetsBalances';
 export { default as useAuction } from './useAuction';
 export { default as useAvailableToSoloStake } from './useAvailableToSoloStake';
 export { default as useBalances } from './useBalances';
+export { default as useBalancesOnAssethub } from './useBalancesOnAssethub';
+export { default as useBalancesOnMultiAssetChain } from './useBalancesOnMultiAssetChain';
+export { default as useBalancesOnSingleAssetChain } from './useBalancesOnSingleAssetChain';
 export { default as useBlockInterval } from './useBlockInterval';
 export { default as useCanPayFee } from './useCanPayFee';
 export { default as useCanPayFeeAndDeposit } from './useCanPayFeeAndDeposit';
@@ -75,6 +78,7 @@ export { default as usePendingRewards } from './usePendingRewards';
 export { default as usePendingRewards2 } from './usePendingRewards2';
 export { default as usePeopleChain } from './usePeopleChain';
 export { default as usePool } from './usePool';
+export { default as usePoolBalances } from './usePoolBalances';
 export { default as usePoolConsts } from './usePoolConsts';
 export { usePoolMembers } from './usePoolMembers';
 export { default as usePools } from './usePools';
