+++ packages/extension-polkagate/src/hooks/useBalancesOnSingleAssetChain.ts
@@ -1,0 +1,144 @@
+// Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+import type React from 'react';
+//@ts-ignore
+import type { FrameSystemAccountInfo } from '@polkadot/types/lookup';
+import type { BalancesInfo, SavedBalances } from '../util/types';
+
+import { useCallback, useContext, useEffect, useState } from 'react';
+
+import { BN, BN_ZERO } from '@polkadot/util';
+
+import { FetchingContext } from '../components';
+import { NATIVE_TOKEN_ASSET_ID } from '../util/constants';
+import { useInfo, useStakingAccount } from '.';
+
+export default function useBalancesOnSingleAssetChain (address: string | undefined, refresh?: boolean, setRefresh?: React.Dispatch<React.SetStateAction<boolean>>, onlyNew = false): BalancesInfo | undefined {
+  const stakingAccount = useStakingAccount(address);
+  const { account, api, chain, chainName, decimal: currentDecimal, formatted, token: currentToken } = useInfo(address);
+  const isFetching = useContext(FetchingContext);
+
+  const [balances, setBalances] = useState<BalancesInfo | undefined>();
+  const [newBalances, setNewBalances] = useState<BalancesInfo | undefined>();
+
+  const token = api?.registry.chainTokens[0];
+  const decimal = api?.registry.chainDecimals[0];
+
+  const getBalances = useCallback(() => {
+    if (!chainName || api?.genesisHash?.toString() !== chain?.genesisHash || !decimal || !token) {
+      return;
+    }
+
+    const ED = api.consts['balances'] ? api.consts['balances']['existentialDeposit'] as unknown as BN : BN_ZERO;
+
+    formatted && api.derive.balances?.all(formatted).then((allBalances) => {
+      //@ts-ignore
+      api.query['system']['account'](formatted).then(({ data: systemBalance }: FrameSystemAccountInfo) => {
+        // some chains such as PARALLEL does not support this call hence BN_ZERO is set for them
+        const frozenBalance = systemBalance?.frozen || BN_ZERO;
+
+        setNewBalances({
+          ED,
+          assetId: NATIVE_TOKEN_ASSET_ID,
+          ...allBalances,
+          chainName,
+          date: Date.now(),
+          decimal,
+          frozenBalance,
+          genesisHash: api.genesisHash.toString(),
+          token
+        });
+        setRefresh && setRefresh(false);
+        isFetching.fetching[String(formatted)]['balances'] = false;
+        isFetching.set(isFetching.fetching);
+      }).catch(console.error);
+    }).catch(console.error);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [api, chain?.genesisHash, chainName, formatted, isFetching.fetching[String(formatted)]?.['length'], setRefresh]);
+
+  useEffect(() => {
+    if (!formatted || !token || !decimal || !chainName || api?.genesisHash?.toString() !== chain?.genesisHash) {
+      return;
+    }
+
+    /** to fetch a formatted address's balance if not already fetching */
+    if (!isFetching.fetching[String(formatted)]?.['balances']) {
+      if (!isFetching.fetching[String(formatted)]) {
+        isFetching.fetching[String(formatted)] = {};
+      }
+
+      isFetching.fetching[String(formatted)]['balances'] = true;
+      isFetching.set(isFetching.fetching);
+      getBalances();
+    } else {
+      console.info(`Balance is fetching for ${formatted}, hence doesn't need to fetch it again!`);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [api, chain?.genesisHash, chainName, decimal, formatted, getBalances, isFetching.fetching[String(formatted)]?.['length'], token]);
+
+  useEffect(() => {
+    if (refresh) {
+      setBalances(undefined);
+      setNewBalances(undefined);
+
+      if (isFetching.fetching[String(formatted)]) {
+        isFetching.fetching[String(formatted)]['balances'] = true;
+      }
+
+      isFetching.set(isFetching.fetching);
+      getBalances();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [Object.keys(isFetching?.fetching ?? {})?.length, formatted, getBalances, refresh]);
+
+  useEffect(() => {
+    if (!chainName || !account || account?.genesisHash !== chain?.genesisHash) {
+      return;
+    }
+
+    // to LOAD saved balances
+    const savedBalances = JSON.parse(account?.balances ?? '{}') as SavedBalances;
+
+    if (savedBalances[chainName]) {
+      const sb = savedBalances[chainName].balances;
+
+      const lastBalances = {
+        ED: new BN(sb['ED'] || '0'),
+        assetId: sb['assetId'] && parseInt(sb['assetId']),
+        availableBalance: new BN(sb['availableBalance']),
+        chainName,
+        date: savedBalances[chainName].date,
+        decimal: savedBalances[chainName].decimal,
+        freeBalance: new BN(sb['freeBalance']),
+        frozenBalance: new BN(sb['frozenBalance'] || '0'),
+        genesisHash: sb['genesisHash'],
+        lockedBalance: new BN(sb['lockedBalance']),
+        pooledBalance: new BN(sb['pooledBalance']),
+        reservedBalance: new BN(sb['reservedBalance']),
+        token: savedBalances[chainName].token,
+        vestedBalance: new BN(sb['vestedBalance']),
+        vestedClaimable: new BN(sb['vestedClaimable']),
+        votingBalance: new BN(sb['votingBalance'])
+      } as BalancesInfo;
+
+      setBalances({
+        ...lastBalances,
+        soloTotal: stakingAccount?.stakingLedger?.total as unknown as BN
+      });
+
+      return;
+    }
+
+    setBalances(undefined);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [Object.keys(account ?? {})?.length, address, chainName, stakingAccount]);
+
+  if (onlyNew) {
+    return newBalances; // returns balances that have been fetched recently and are not from the local storage, and it does not include the pooledBalance
+  }
+
+  return balances && balances.token === currentToken && balances.decimal === currentDecimal
+    ? balances
+    : undefined;
+}
