+++ packages/extension-polkagate/src/hooks/useTokenPrice.ts
@@ -25,13 +25,14 @@ const assetsChains = createAssets();
  * @param address : accounts substrate address
  * @returns price : price of the token which the address is already switched to
  */
-export default function useTokenPrice (address: string | undefined, assetId?: number): Price | typeof DEFAULT_PRICE {
+export default function useTokenPrice (address: string | undefined, assetId?: number |string): Price | typeof DEFAULT_PRICE {
   const { chainName, genesisHash } = useInfo(address);
   const userAddedPriceId = useUserAddedPriceId(genesisHash);
   const pricesInCurrencies = usePrices();
   const mayBeAssetsOnMultiAssetChains = assetsChains[toCamelCase(chainName || '')];
 
   const isAssetHub = ASSET_HUBS.includes(genesisHash || '');
+  const isForeignAsset = isAssetHub && assetId && typeof assetId === 'string' && assetId.startsWith('0x');
 
   const _assetId = assetId !== undefined
     ? assetId
@@ -45,8 +46,8 @@ export default function useTokenPrice (address: string | undefined, assetId?: nu
     }
 
     // FixMe, on second fetch of asset id its type will get string which is weird!!
-    const priceId = _assetId !== undefined && _assetId > NATIVE_TOKEN_ASSET_ID
-      ? mayBeAssetsOnMultiAssetChains?.find(({ id }) => id === Number(_assetId))?.priceId
+    const priceId = _assetId !== undefined && (_assetId as number > NATIVE_TOKEN_ASSET_ID || isForeignAsset)
+      ? mayBeAssetsOnMultiAssetChains?.find(({ id }) => id === Number(_assetId) || id === _assetId)?.priceId
       : userAddedPriceId || getPriceIdByChainName(chainName);
 
     const mayBePriceValue = priceId ? pricesInCurrencies.prices?.[priceId]?.value || 0 : 0;
@@ -56,5 +57,5 @@ export default function useTokenPrice (address: string | undefined, assetId?: nu
       priceChainName: chainName?.toLocaleLowerCase(),
       priceDate: pricesInCurrencies.date
     };
-  }, [_assetId, chainName, mayBeAssetsOnMultiAssetChains, pricesInCurrencies, userAddedPriceId]);
+  }, [_assetId, chainName, isForeignAsset, mayBeAssetsOnMultiAssetChains, pricesInCurrencies, userAddedPriceId]);
 }
