+++ packages/extension-polkagate/src/util/workers/getAssetOnAssetHub.js
@@ -1,44 +1,35 @@
 // Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-/* eslint-disable import-newlines/enforce */
-/* eslint-disable object-curly-newline */
-
 import { BN_ZERO } from '@polkadot/util';
 
-// eslint-disable-next-line import/extensions
+import { decodeMultiLocation } from '../utils';
 import { closeWebsockets, fastestEndpoint, getChainEndpoints, metadataFromApi, toGetNativeToken } from './utils';
 
-// @ts-ignore
-async function getAssetOnAssetHub (addresses, assetsToBeFetched, chainName, userAddedEndpoints) {
-  const endpoints = getChainEndpoints(chainName, userAddedEndpoints);
-  const { api, connections } = await fastestEndpoint(endpoints);
-
-  const result = metadataFromApi(api);
-
-  postMessage(JSON.stringify(result));
-
-  const results = await toGetNativeToken(addresses, api, chainName);
-
-  // @ts-ignore
-  const nonNativeAssets = assetsToBeFetched.filter((asset) => !asset.extras?.isNative);
-
-  for (const asset of nonNativeAssets) {
+//@ts-ignore
+async function getAssets (addresses, api, assets, chainName, results) {
+  for (const asset of assets) {
+    const isForeignAssets = asset.isForeign;
+    const section = isForeignAssets ? 'foreignAssets' : 'assets';
+    const assetId = isForeignAssets ? decodeMultiLocation(asset.id) : asset.id;
     // @ts-ignore
-    const maybeTheAssetOfAddresses = addresses.map((address) => api.query.assets.account(asset.id, address));
-    const assetMetaData = api.query.assets.metadata(asset.id);
+    const maybeTheAssetOfAddresses = addresses.map((address) => api.query[section].account(assetId, address));
+    const assetMetaData = api.query[section].metadata(assetId);
 
     const response = await Promise.all([assetMetaData, ...maybeTheAssetOfAddresses]);
     const metadata = response[0];
-    const AssetOfAddresses = response.slice(1);
+    const assetOfAddresses = response.slice(1);
 
     const decimal = metadata.decimals.toNumber();
     const token = metadata.symbol.toHuman();
 
-    AssetOfAddresses.forEach((_asset, index) => {
+    assetOfAddresses.forEach((_asset, index) => {
       const balance = _asset.isNone ? BN_ZERO : _asset.unwrap().balance;
-      const parsedAccountAsset = JSON.parse(JSON.stringify(_asset));
-      const isFrozen = parsedAccountAsset?.status === 'Frozen';
+
+      const isFrozen = isForeignAssets
+        ? metadata.isFrozen.valueOf()
+        : _asset.isSome && _asset.unwrap().status.valueOf().toString() === 'Frozen';
+
       const _balance = String(balance);
 
       const item = {
@@ -54,6 +45,7 @@ async function getAssetOnAssetHub (addresses, assetsToBeFetched, chainName, user
         frozenBalance: isFrozen ? balance : 0, // JUST to comply with the rule that total=available + reserve
         genesisHash: api.genesisHash.toString(),
         isAsset: true,
+        isForeignAssets: !!isForeignAssets,
         priceId: asset?.priceId,
         token,
         totalBalance: _balance
@@ -65,6 +57,34 @@ async function getAssetOnAssetHub (addresses, assetsToBeFetched, chainName, user
       results[_index]?.push(item) ?? (results[_index] = [item]);
     });
   }
+}
+
+// @ts-ignore
+async function getAssetOnAssetHub (addresses, assetsToBeFetched, chainName, userAddedEndpoints) {
+  const endpoints = getChainEndpoints(chainName, userAddedEndpoints);
+  const { api, connections } = await fastestEndpoint(endpoints);
+
+  const result = metadataFromApi(api);
+
+  postMessage(JSON.stringify(result));
+
+  const results = await toGetNativeToken(addresses, api, chainName);
+
+  // @ts-ignore
+  const nonNativeAssets = assetsToBeFetched.filter((asset) => !asset.extras?.isNative);
+
+  /** to calculate a new Foreign Token like MYTH asset id based on its XCM multi-location */
+  // const allForeignAssets = await api.query.foreignAssets.asset.entries();
+  // for (const [key, _others] of allForeignAssets) {
+  //   const id = key.args[0];
+  //   const assetMetaData = await api.query.foreignAssets.metadata(id);
+
+  //   if (assetMetaData.toHuman().symbol === 'MYTH') {
+  //     console.log('new foreign asset id:', encodeMultiLocation(id));
+  //   }
+  // }
+
+  await getAssets(addresses, api, nonNativeAssets, chainName, results);
 
   postMessage(JSON.stringify(results));
   closeWebsockets(connections);
